[frontend] refactor render card title with country flag
